requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
drf-yasg==1.21.7
PyJWT==2.8.0
setuptools>=65.5.1
